Improve shared event loop shutdown coverage and formatting
Move unit tests into ydb/_topic_common/common_test.py so they run under
the unit coverage job, cover recreate/idempotent/pending-task paths, and
apply black formatting required by CI.
